Library

Public API

SuperVFM.GetSchwarzTempCoeffs — Method

GetSchwarzTempCoeffs(Temp::Real)

Obtain the Schwarz α and α' parameters from observational data using the temperature. For values that lie within the observation data, compute a cubic spline interpolation to extract parameters for the desired temperature.

SuperVFM.check_timestep — Method

check_timestep(SimParams::SimulationParams)

Checks if the current timestep in SimParams is small enough to resolve the smallest Kelvin waves. Returns true if the timestep $\Delta t < \Delta t_{max}$.

The maximum timestep is given by

\[ \Delta t_{max} = \frac{(\delta/2)^2}{\kappa\log(\delta/(2\pi a_0))}\]

SuperVFM.get_density — Method

get_density(T::AbstractFloat)

Returns the normal fluid density $ρ_n$ and superfluid density $ρ_s$ for a given temperature $T$ in arbitrary units of temperature.

Requires $0.0\leq T \leq T_{\lambda} = 2.178$

SuperVFM.get_density — Method

get_density(T::Unitful.Temperature)

Returns the normal fluid density $ρ_n$ and superfluid density $ρ_s$ for a given temperature $T$ in arbitrary units of temperature.

Requires $0.0\leq T \leq T_{\lambda} = 2.178$

SuperVFM.get_dynamic_viscosity — Method

get_dynamic_viscosity(T::AbstractFloat)

Returns the dynamic viscosity $η$ at temperature $T$.

Requires $0.8\leq T \leq T_{\lambda} = 2.178$

SuperVFM.get_dynamic_viscosity — Method

get_dynamic_viscosity(T::Unitful.Temperature)

Returns the dynamic viscosity $η$ at temperature $T$.

Requires $0.8\leq T \leq T_{\lambda} = 2.178$

SuperVFM.print_characteristics — Method

print_characteristics(io::IO,SimParams::SimulationParams)

Prints the characteristic time and length scales of the simulation in dimensional units

SuperVFM.print_density_data — Function

print_density_data(io::IO=stdout)

Prints the normal fluid $ρ_n$ and superfluid $ρ_s$densities to the IO buffer. Defaults to stdout.

SuperVFM.print_dynamic_visosity_data — Function

print_dynamic_visosity_data(io::IO=stdout)

Prints the dynamic viscosity η to the IO buffer. Defaults to stdout.

SuperVFM.LIA — Method

(Velocity::LIA)(f, ghosti, ghostb, Empty, κ, corea)

Computes the superfluid velocity using the Local Induction Approximation (LIA)

\[\mathbf{v}_i = \beta \mathbf{s}'_i \times \mathbf{s}''_i \]

SuperVFM.OpenBoundary — Method

OpenBoundary(dim::Int)

Initialises an open boundary in the chosen direction. Vortex loops that exceed the box size (typically $2π$) are not restricted. Set dim=1 for the $x$ direction, dim=2 for the $y$ direction and dim=3 fir the $z$ direction.

Example usage:

    boundary_x = OpenBoundary(1)
    boundary_y = OpenBoundary(2)
    boundary_z = OpenBoundary(3)

SuperVFM.PeriodicBoundary — Method

PeriodicBoundary(dim::Int)

Initialises a periodic boundary in the chosen direction selected by dim. Vortex loops that exceed the box size (typically $2π$) are looped back periodically to the other side of the box.Set dim=1 for the $x$ direction, dim=2 for the $y$ direction and dim=3 fir the $z$ direction.

Example usage:

    boundary_x = PeriodicBoundary(1)
    boundary_y = PeriodicBoundary(2)
    boundary_z = PeriodicBoundary(3)

Warning

As of v1.0.2, only the periodic boundary condition is fully implemented and working correctly, open boundary conditions and solid walls will be implemented in a future release.

Internal API

Base.show — Method

Base.show(io::IO,SimParams::SimulationParams)

Lists the simulation parameters stored in SimParams in a stylistic way with a key.

SuperVFM.get_deriv_1 — Method

get_deriv_1(f::AbstractArray, ghosti::AbstractArray, ghostb::AbstractArray, Empty::Bool)

Computes the first order derivative using a 2nd order finite difference adaptive scheme using ghost points, skipping filaments that are labelled as empty.

SuperVFM.get_deriv_2 — Method

get_deriv_2(f::AbstractArray, ghosti::AbstractArray, ghostb::AbstractArray, Empty::Bool)

Computes the first order derivative using a 2nd order finite difference adaptive scheme, skipping filaments that are labelled as empty.